[scanner] fix: split highest-hook-count mission chat components
Decompose MissionChatView.tsx (546 lines / 60 hooks) into concern-based
custom hooks: useMissionChatScroll, useMissionChatComposer,
useMissionChatTitle, useMissionChatMissionEditor, useMissionChatFeedback,
useMissionChatLifecycle, and useMissionChatDerived. The container is now
297 lines with 13 hook calls (both under the <300 line / <15 hook target),
with no behaviour change — data-fetching stays in the container and props
are passed down unchanged.

Fixes #21509
Fixes #21628
