+++ web/src/components/layout/mission-sidebar/mission-chat/missionChatConstants.ts
@@ -2,6 +2,8 @@
 export const SCROLL_BOTTOM_THRESHOLD_PX = 50
 /** Duration in ms for the scroll-to-bottom button fade animation */
 export const SCROLL_BTN_FADE_MS = 200
+/** Delay in ms before auto-scrolling to the bottom when entering full-screen */
+export const FULLSCREEN_SCROLL_DELAY_MS = 100
 export const COMPACT_ACTION_BUTTON_CLASS = 'h-10 w-10 shrink-0 p-0 inline-flex items-center justify-center rounded-lg'
 export const MISSION_PROGRESS_MIN = 0
 export const MISSION_PROGRESS_MAX = 100
